+++ examples/futures/Trading/Claim.php
@@ -1,0 +1,11 @@
+<?php
+
+use BitMart\Futures\APIContractTrading;
+require_once __DIR__ . '/../../ExampleConfig.php';
+require_once __DIR__ . '/../../../vendor/autoload.php';
+
+$APIContract = new APIContractTrading(ExampleConfig::getExampleConfig());
+
+// Claim demo trading assets
+$response = $APIContract->claim()['response'];
+echo json_encode($response);